auth_api/python/requirements.txt
Flask
mysql-connector-python
PyJWT